Phase C-1 cascade: move PendingFieldInfo + PendingFieldLayoutSummary
resume.py pending heap-write deferred-replay metadata:
- PendingFieldLayoutSummary (serialization-friendly summary) →
  majit-ir::resumedata (deps: DescrRef + ResumeValueKind +
  ResumeValueLayoutSummary, all already in majit-ir).
- PendingFieldInfo (live ResumeValueSource-bearing entry) →
  majit-backend::resume_value (deps: DescrRef + ResumeValueSource).
- opt_descr_arc_ptr_eq helper (history.py:125 Arc::ptr_eq parity for
  Option<DescrRef>) → majit-ir::resumedata.

PendingFieldLayoutSummary inherent impls (to_pending_field_info /
to_exit_pending_field_layout) reference metainterp/backend boundary
types and convert through ResumeValueLayoutSummaryExt; extracted as
free functions pending_field_layout_to_pending_field_info /
pending_field_layout_to_exit_pending_field_layout in metainterp.

Re-exported from majit-metainterp::resume for caller compatibility.
